Create DrillsAndCutters.com-Scrape
My code for scraping DrillsAndCutters.com for product information. Written with Python, Selenium, and BeautifulSoup.

I work with a web development agency that services the industrial and construction industry. Some of our high-profile clients include NewSouthConstructionSupply.com and NTRuddock.com. We gather information from manufacturers and, if a client wishes to sell that manufacturer's products on their website, we'll scrape the manufacturer's website for product information.

This benefits both the manufacturer and our clients, and increases sales for both parties.

Please use scraping responsibly.